"Added geosynchronous orbit calculations and improved 'IsCircular' check
This update introduces the ability to calculate geosynchronous orbits with the 'GeosynchronousOrbit' method in the 'CelestialBody.cs' file. Additionally, the 'IsCircular' method in 'OrbitalParameters.cs' has been modified for better accuracy, now considering orbits with an eccentricity less than 1E-03 to be circular. To accommodate these changes, corresponding test cases have also been added or updated in 'CelestialBodyTests.cs' and 'KeplerianElementsTests.cs'. This feature is essential for various celestial physics calculations and for the practical application of launching and positioning satellites in geosynchronous orbits."