Feature/heliosynchronous (#67)
* "Added geosynchronous orbit calculations and improved 'IsCircular' check

This update introduces the ability to calculate geosynchronous orbits with the 'GeosynchronousOrbit' method in the 'CelestialBody.cs' file. Additionally, the 'IsCircular' method in 'OrbitalParameters.cs' has been modified for better accuracy, now considering orbits with an eccentricity less than 1E-03 to be circular. To accommodate these changes, corresponding test cases have also been added or updated in 'CelestialBodyTests.cs' and 'KeplerianElementsTests.cs'. This feature is essential for various celestial physics calculations and for the practical application of launching and positioning satellites in geosynchronous orbits."

* Add HelioSynchronousOrbit and TrueSolarDay calculations in CelestialBody

The CelestialBody class was expanded to handle calculations for heliosynchronous orbit and true solar day. They were added to provide more orbital options and accurate scientific computations for solar days for the respective celestial bodies. J2, J3, and J4 properties were added to CelestialBody and DTO.CelestialBody classes to store the second, third, and fourth zonal harmonic coefficients respectively from the gravitational potential, essential for orbital mechanics. The Plane class was created to assist these calculations, and tests were updated accordingly. Adjustments were made in OrbitalParameters for consistency.

* Update accuracy of test assertions and refactor celestial calculations

The assertions in `CelestialBodyTests.cs` have been modified to include precision parameters to enhance accuracy of the tests. Additionally, a refactor has taken place in `CelestialBody.cs` for the calculations regarding ephemeris and mean anomalies. This refactoring involved updating formulas and adjusting calling parameters to improve the accuracy and reliability of the celestial calculations.

* Add phased heliosynchronous orbit support in CelestialBody class

Expanded the CelestialBody class to support calculations for a phased heliosynchronous orbit through methods and tests. The expanded class now allows computing for the Keplerian elements of a phased heliosynchronous orbit. Corresponding test cases were also added in CelestialBodyTests.cs to guarantee the functionality of these new computational methods.

* Add VectorComparer method for testing in Astrodynamics

A new VectorComparer method has been added to the Astrodynamics unit tests, providing a reliable technique for comparing vector values with a precision of 1E-03. The method is thread-safe using a lock and has been incorporated into the CelestialBodyTests.

---------

Co-authored-by: Sylvain Guillet <s.guillet@quantumsurgical.com>