Updated Math and Maneuver classes, added new methods
This update includes significant changes to classes in Astrodynamics/Math and Astrodynamics/Maneuver. In the Math group, Plane and Vector3 classes have been upgraded with new methods. In the Maneuver group, many classes added a new `ManeuverPointComputation` method, which currently throws NotImplementedException. The ComputeDeltaV, ComputeDeltaT, and ComputeDeltaM methods are also updated in the Maneuver class.