Add performance testing for spacecraft propagation
The commit includes the addition of the BenchmarkDotNet library to the IO.Astrodynamics.Tests project. A new 'Performance' class is created in the same project, with a benchmark test for the 'ScenarioTests' class. Additionally, a new project 'IO.Astrodynamics.Performance' has been added to the solution for dedicated performance testing.